requirements.txt
Flask~=3.1.1
Flask-Login~=0.6.3
SQLAlchemy~=2.0.41
WTForms~=3.2.1
Flask-Migrate~=4.1.0
Werkzeug~=3.1.3
sentry-sdk~=2.32.0
structlog~=25.4.0
Flask-Caching~=2.3.1
Flask-Limiter~=3.12
flask-talisman~=1.1.0
python-dotenv~=1.1.1